Dont fold over ir files
